pyproject.toml
[project]
name = "bili-hardcore"
version = "2.0.0"
description = "B站硬核会员自动答题工具 - WebUI 版"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "httpx>=0.28.0",
    "aiosqlite>=0.21.0",
    "python-jose[cryptography]>=3.3.0",

    "python-multipart>=0.0.18",
    "qrcode[pil]>=8.0",
    "pillow>=11.0.0",
]

[project.scripts]
bili-hardcore = "bili_hardcore.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bili_hardcore"]
